fix(emoji): clean up uploaded blob when create/replace is rejected
The image is streamed to storage before the command validates, so a rejected
create (duplicate, 409) or replace (missing, 404) left the blob orphaned in
storage with nothing referencing it. Route uploads now go through a shared
saveEmojiUpload helper that removes the blob if the command throws. Expose
list() on the storage facade and assert no leak in the 409/404 tests.
